Update sqlcmd-download-install.md
It's VERY frustrating that a trillion dollar company cares so little about its users that when we TRY to push fixes to your broken documentation, you lose our changes when your site has issues. It is not amusing to see a white screen when a huge unicorn image when this happens. Maybe spend less time trying to be cute ant more time making for a robust system. Even better, allow issues to be created in this repo so that we don't have to do the work of MS employees.

Today our deployment pipelines started failing while trying to install go-sqlcmd on the deployment agent machine, getting error `Unable to handle repository shortcut 'deb [arch=amd64,arm64,armhf signed-by=/usr/share/keyrings/microsoft-prod.gpg] https://packages.microsoft.com/ubuntu/22.04/prod jammy main'`. 

At this point I'm just quoting copilot because it is so accurate:

```
The failure is due to a recent change in how Microsoft publishes their APT repository instructions. The new deb ... signed-by=... format is not supported by add-apt-repository—it expects a simple deb ... line without the signed-by option. This causes the error you see.

How to fix:
Instead of using add-apt-repository with the new format, write the repository line directly to a .list file in /etc/apt/sources.list.d/. This is the recommended approach for modern Microsoft package repos.

...

Summary of changes:

- Imports the Microsoft GPG key in the new recommended location.
- Writes the repo line directly to /etc/apt/sources.list.d/microsoft-prod.list.
- Removes add-apt-repository entirely.
```

Note: I have not yet tested this. I'll comment in a few minutes if it goes well.